# data-layer/database/mongo-schemas/mongo-refresh-token.js
const mongoose = require("mongoose")

const refreshTokenSchema = new mongoose.Schema({
    token: {
        type: String,
        required: true,
    },
})

module.exports = mongoose.model("RefreshToken", refreshTokenSchema)